Add files via upload
This is an altered implementation of the El Farol Bar problem. The original problem describes a situation on a Thursday night at the El Farol Bar in New Mexico, USA. Every Thursday night there is Irish Music playing. 100 people want to attend, but the bar can only hold 60 people. Agents try and occupy the bar with only information back the last Thursdays occupation. They therefore have to make a prediction on whether or not to go to the bar without any other information.   
This altered problem that I have implemented takes the original idea and adds multiple bars. Multiple Agents then try and learn how to optimally occupy the bar on each night learning via a defined reward.